chore: repo setup incl. CI, license, etc. (#3)
The `go.yml` and `lint.yml` CI workflows come from `libevm` as
reasonable first approximations of what we want for this repo.

The failures on the Go workflow are all to be expected because there
aren't any Go files to test, generate, nor lint. What's important is
that all workflows are running.