fix: remove `golangci-lint` config only needed by `libevm` (#4)
`goheader` only checks new issues in `libevm` because `MOD-YEAR-RANGE`
breaks otherwise (we can reintroduce this next year if we haven't moved
the code into `avalanchego` yet). The other deleted config is what
`libevm` uses to avoid linting `geth` code.

Go workflows are still expected to fail, as described in #3, and all
that matters is that they run (and `golangci-lint` _attempts_ to lint
but finds no files).