feat: `adaptor` package (#7)
Allows block implementations to be read-only and not have to be aware of
the VM's that created them, which is typically necessary when calling
`Block.Verify()`, `Accept()`, or `Reject()`. These methods instead exist
on the VM as `{Verify,Accept,Reject}Block()`.

There are no explicit tests for this package as it's exercised heavily
via the actual SAE tests that will follow.