feat: `blocks` package (#18)
Introduces the `blocks.Block` type (and associated helpers), which
satisfies the `adaptor.BlockProperties` interface introduced in #7.
While the motivation of specific functionality may not be clear in
isolation, context can be garnered from the prototype code (f79aff08)
and how it uses this package in `saexec` and the primary SAE
implementation.

---------

Signed-off-by: Arran Schlosberg <519948+ARR4N@users.noreply.github.com>
Co-authored-by: Stephen Buttolph <stephen@avalabs.org>