feat: `blockstest` and `saetest.Wallet` test helpers (#24)
Introduces test-only builders of blocks and transactions. Maintaining
block lineage and per-key nonces is a common pattern that I noticed in
all of my SAE development, so these abstract them.

Testing is performed against a `libevm/core.BlockChain` as the gold
standard, as these helpers will be key to testing the `saexec.Executor`
to be introduced later.

---------

Signed-off-by: Arran Schlosberg <519948+ARR4N@users.noreply.github.com>
Co-authored-by: Austin Larson <78000745+alarso16@users.noreply.github.com>