Remove hook.BeforeBlock and hook.AfterBlock (#40)
In implementing the worstcase gas pricing, I want to share the gastime
updating logic as much as possible. However, that should not result in
calling `hook.Points.BeforeExecutingBlock` or
`hook.Points.AfterExecutingBlock`.

This PR separates the gastime changes and also includes fuzzing to
ensure that updating the target can not invalidate the worst case gas
price assumptions.

---------

Signed-off-by: Stephen Buttolph <stephen@avalabs.org>
Co-authored-by: Arran Schlosberg <519948+ARR4N@users.noreply.github.com>