test: require fuzz corpora but don't automatically fuzz (#46)
Each fuzz target has different requirements for the duration so in some
cases the 15m was wasteful while in others it was insufficient. I've
instead changed to requiring the inclusion of corpora (which are run
with every call to `go test`, not just those with the `-fuzz` flag) and
provided a `Taskfile` to build and copy them locally.

This approach places the onus on the author of the fuzz test to decide
on a reasonable amount of coverage discovery[^1]. We can always add
ongoing fuzzing later, but I don't think it's terribly important right
now.

[^1]: For example, note the much larger corpus for `FuzzOpCodes`
relative to `FuzzEffectiveGasTip`.