Add EndOfBlockOps hook (#47)
AtomicTxs are separate transactions (completely unrelated to the
ethereum transactions) which are stored in a libevm extension to the
block body.

The AtomicTxs have the ability to spend native tokens from an EOA and to
mint new tokens from cross-chain transfers. Additionally, they include
multi-coin transfers, which will be implemented by the
`AfterExecutingBlock` callback.

Each `Op` is going to correspond to a single atomic tx, which is why a
slice is returned.