refactor: abstract `safeMaxBlockSize()` and introduce concept of open/closed queue
The thresholds of the SAE queue are specialised and require precise knowledge of the ACP to understand.
This commit aims to remove that knowledge burden and provide to the reader directly.

I abstracted the function to remove any possibility (in code or when read) that the max rate modified the gas clock.
My first thought when reading the code was "why are we changing the gas rate?" so I wanted to demonstrate that we're not.