Add `worstcase.State` (#41)
This PR introduces an initial pass of worst-case state metering.

Worstcase state is intended to be used during block building (and
verification) to ensure that the block can be included based on the
queue, as well as ensuring that all included transactions will be able
to be executed according to the EVM rules.

---------

Signed-off-by: Stephen Buttolph <stephen@avalabs.org>
Co-authored-by: Arran Schlosberg <519948+ARR4N@users.noreply.github.com>