feat: `VM.AcceptBlock()` implementation w/ invariant documentation (#59)
The majority of this PR is documentation that defines the order in which
side effects of the block lifecycle MUST occur. The implementation of
`VM.AcceptBlock()` (which, by definition, includes block settlement)
necessitated these ordering invariants to avoid race conditions in
current and future code.

As the block lifecycle is now (roughly) complete, I thought the most
valuable test would be integration rather than unit. The PR therefore
includes the necessary (and concise) RPC-backend methods to support
`eth_getHeaderByNumber` as it exercises block labels as well as
canonical blocks.

---------

Signed-off-by: Arran Schlosberg <519948+ARR4N@users.noreply.github.com>
Co-authored-by: Stephen Buttolph <stephen@avalabs.org>